pyproject.toml
[tool.poetry]
name = "carona-parque"
version = "0.1.0"
description = "My own template for Python projects."
authors = ["Gabriel Gazola Milan <gabriel.gazola@poli.ufrj.br>"]
license = "GPL-3.0-only"
readme = "README.md"
homepage = "https://github.com/gabriel-milan/carona-parque"
repository = "https://github.com/gabriel-milan/carona-parque"
keywords = ["python", "template"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.4"
celery = "^5.2.6"
loguru = "^0.6.0"
python-telegram-bot = "^13.12"
psycopg2-binary = "^2.9.3"
django-celery-beat = "^2.3.0"
redis = "^4.3.3"
gunicorn = "^20.1.0"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
pre-commit = "^2.18.1"
pytest-cov = "^3.0.0"
flake8 = "^4.0.1"
pdoc3 = "^0.10.0"

[tool.poetry.scripts]
docs = "scripts.docs:main"
lint = "scripts.lint:main"
test = "scripts.test:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
